src/pyproject.toml
[tool.poetry]
name = "django-n-postgres template"
version = "0.1.0"
description = ""
authors = []
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "5.0.8"
djangorestframework = "3.15.1"
gunicorn = "22.0.0"
psycopg = "3.1.19"
psycopg-binary = "3.1.19"
celery = "5.4.0"
django-celery-results = "2.5.1"
amqp = "5.2.0"
json-log-formatter = "1.0"
dependency-injector = "4.41.0"
concurrent-log-handler = "0.9.25"
drf-spectacular = "0.27.2"
django-stubs = "5.0.4"

[tool.poetry.group.dev.dependencies]
pytest = "8.2.1"
pytest-django = "4.8.0"
pytest-cov = "5.0.0"
pytest-mock = "3.14.0"
pytest-timeout = "2.3.1"
mypy = "1.10.0"
flake8 = "7.1.0"
black = "24.4.2"
pre-commit = "3.7.1"
isort = "5.13.2"

[tool.poetry.group.test.dependencies]
pytest = "8.2.1"
pytest-django = "4.8.0"
pytest-cov = "5.0.0"
pytest-mock = "3.14.0"
pytest-timeout = "2.3.1"

[tool.poetry.group.lint.dependencies]
flake8 = "7.1.0"

[tool.poetry.group.analyze.dependencies]
mypy = "1.10.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = "test_*.py"
addopts = [
    "-v",
    "-rfxP",
    "--cov",
    "--log-level=CRITICAL",
]

[tool.coverage.report]
exclude_also = [
    # Don't complain about missing debug-only code:
    "def __repr__",
    "if self\\.debug",

    # Don't complain if tests don't hit defensive assertion code:
    "raise AssertionError",
    "raise NotImplementedError",

    # Don't complain if non-runnable code isn't run:
    "if 0:",
    "if __name__ == .__main__.:",

    # Don't complain about abstract methods, they aren't run:
    "@(abc\\.)?abstractmethod",
]
omit = [
    "manage.py",
    "*conftest.py",
    "*models.py",
    "*apps.py",
    "*urls.py",
    "*schema.py",
    "*admin.py",
    "*forms.py",
    "*filters.py",
    "*interfaces.py",
    "*utils/tests.py",
    "*utils/serializers.py",
    "*utils/pagination.py",
    "*utils/openapi_examples.py",
    "*utils/logging.py",
    "*utils/exceptions.py",
    "*utils/admin.py",
    "*__init__*",
    "*/config/*",
    "*/migrations/*",
    "*/tests/*",
    "*/views/*",
]

[tool.black]
exclude = '''
    migrations
    | .venv
'''

[tool.isort]
profile = "black"
skip_glob = ["di/*"]

[tool.mypy]
python_version = 3.11
exclude = [
    "migrations",
    "views",
    "tests",
    "admin\\.py",
    "utils\\.py",
    "schema\\.py",
    "apps\\.py",
    "tests\\.py",
    "forms\\.py",
    "admin\\.py",
    "dev\\.py",
    "test\\.py",
]
ignore_missing_imports = true
disable_error_code = "import-untyped"
follow_imports = "skip"
plugins = [
    "mypy_django_plugin.main",
]

[tool.django-stubs]
django_settings_module = "config.settings"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
